GKAAA0 passes but only once
basic still prints 1. as -1
the problem is only in the conversion. if all arguments are integers,
the computations are correct. e.g. print 1/10 or sin(2/3) work fine.

wonder why the tests only complete one cycle. maybe something to do with
traps/halt shenanigans, which this emulator barely implements at all.